made the editor more easy to change the colors to by adding global variables. Restructured the editor page to make it a better UI/UX. Changed the accent color temporarily.
