+++ src/app/globals.css
@@ -3,6 +3,7 @@
     --doc-height: 100%;
     --nav-height: 65px;
     --editor-titlebar-height: 40px;
+    --editor-subtitlebar-height: 35px;
     --editor-side-menu-width: 60px;
     --border-radius: 25px;
     --border-radius-small: 10px;
@@ -36,16 +37,23 @@
     --accent-pink: 184, 170, 219;
     --text-color-rgb: 41, 41, 41;
     /* Editor */
+    --editor-accent-color-rgb: 184, 80, 66;
+    --editor-accent-color-hover-rgb: 220, 120, 100;
+    --editor-text-color-rgb: 0, 0, 0;
+    --editor-background-color-rgb: 48, 48, 69;
+    --editor-text-contrast-color-rgb: 255, 255, 255;
     --editor-side-menu-closed-background-color-rgb: 255, 255, 255;
     --editor-side-menu-closed-text-color-rgb: 41, 41, 41;
     --editor-side-menu-background-rgb: 48, 48, 69;
     --editor-side-menu-text-rgb: 200, 200, 200;
-    --editor-side-menu-accent-rgb: 116, 103, 255;
-    --editor-side-menu-accent-hover-rgb: 164, 147, 255;
+    --editor-side-menu-accent-rgb: var(--editor-accent-color-rgb);
+    --editor-side-menu-accent-hover-rgb: var(--editor-accent-color-hover-rgb);
     --editor-side-menu-contrast-background-rgb: 255, 255, 255;
     --editor-side-menu-contrast-text-rgb: 0, 0, 0;
     --editor-side-menu-low-contrast-text-rgb: 150, 150, 150;
     --editor-side-menu-low-contrast-background-rgb: 200, 200, 200;
+    --editor-subtitlebar-background-background-color-rgb: 255, 255, 255;
+    --editor-subtitlebar-button-hover-background-color-rgb: 227, 227, 227;
 }
 
 * {
