+++ src/components/icons/iconSVG.tsx
@@ -306,8 +306,9 @@ export const cancelIcon = (
 );
 
 export const saveIcon = (
-    <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 16 16">
-        <path d="M5 2v2.5A1.5 1.5 0 0 0 6.5 6h2A1.5 1.5 0 0 0 10 4.5V2h.379a2 2 0 0 1 1.414.586l1.621 1.621A2 2 0 0 1 14 5.621V12a2 2 0 0 1-2 2V9.5A1.5 1.5 0 0 0 10.5 8h-5C4.673 8 4 8.669 4 9.498V14a2 2 0 0 1-2-2V4a2 2 0 0 1 2-2zm1 0v2.5a.5.5 0 0 0 .5.5h2a.5.5 0 0 0 .5-.5V2zM5 14h6V9.5a.5.5 0 0 0-.5-.5h-5c-.277 0-.5.223-.5.498z"></path>
+    <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24">
+        <path d="M18.437 20.937H5.563a2.372 2.372 0 0 1-2.5-2.211v-11a2.372 2.372 0 0 1 2.5-2.212h.462a.5.5 0 0 1 0 1h-.462a1.381 1.381 0 0 0-1.5 1.212v11a1.38 1.38 0 0 0 1.5 1.211h12.874a1.38 1.38 0 0 0 1.5-1.211v-11a1.381 1.381 0 0 0-1.5-1.212h-.462a.5.5 0 0 1 0-1h.462a2.372 2.372 0 0 1 2.5 2.212v11a2.372 2.372 0 0 1-2.5 2.211"></path>
+        <path d="m8.645 6.213l3-3a.5.5 0 0 1 .35-.15a.508.508 0 0 1 .36.15l3 3a.5.5 0 0 1-.71.71l-2.14-2.14v8.47a.508.508 0 0 1-.5.5a.5.5 0 0 1-.5-.5v-8.49l-2.15 2.16a.5.5 0 0 1-.71-.71"></path>
     </svg>
 );
