+++ src/features/editor/components/page-util-bar/PageUtilBar.module.css
@@ -5,77 +5,13 @@
     margin-inline: auto;
     margin-block-end: var(--padding-block-small);
     display: flex;
-    justify-content: space-between;
+    justify-content: flex-end;
     align-items: center;
     gap: var(--gap-small);
 }
 
 .pageCounter {
     font-size: var(--font-size-small);
-}
-
-.pageFunctionContainer {
-    display: flex;
-    gap: var(--gap-small);
-    justify-content: center;
-    align-items: center;
-}
-
-.functionButtonContainer {
-    display: flex;
-    gap: var(--gap-medium);
-    align-items: center;
-    justify-content: center;
-    background-color: #7367ff;
-    padding-inline: var(--padding-block-small);
-    padding-block: var(--padding-block-small);
-    border-radius: var(--border-radius-xsmall);
-}
-
-.deletePageButton {
-    background-color: transparent;
-    border: none;
-    width: 15px;
-    height: 15px;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    cursor: pointer;
-}
-
-.deletePageButton svg {
-    width: 100%;
-    height: 100%;
-    stroke: white;
-    fill: white;
-}
-
-.deletePageButton svg:hover {
-    fill: #a4a0d1;
-    stroke: #a4a0d1;
-}
-
-.addPageButton {
-    background-color: transparent;
-    border: none;
-    width: 15px;
-    height: 15px;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    cursor: pointer;
-}
-
-.addPageButton svg {
-    stroke: white;
-    fill: white;
-    width: 100%;
-    height: 100%;
-}
-
-.addPageButton svg:hover {
-    fill: #a4a0d1;
-    stroke: #a4a0d1;
 }
 
 @media screen and (max-width: 768px) {
