+++ src/features/editor/components/title-bar/TitleBar.module.css
@@ -1,11 +1,11 @@
 .container {
     height: var(--editor-titlebar-height);
     width: 100%;
-    background-color: #7367ff;
+    background-color: rgb(var(--editor-accent-color-rgb));
     display: flex;
     flex-direction: row;
     align-items: center;
-    justify-content: space-between;
+    justify-content: center;
     padding-inline: var(--padding-inline-medium);
     color: white;
     gap: var(--gap-medium);
@@ -13,56 +13,4 @@
 
 .title {
     color: white;
-}
-
-.iconContainer {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    gap: var(--gap-small);
-    cursor: pointer;
-}
-
-.iconContainer svg:hover {
-    fill: rgba(239, 239, 239, 0.8);
-    stroke: rgba(239, 239, 239, 0.8);
-}
-
-.iconContainer svg {
-    width: 20px;
-    height: 20px;
-    fill: white;
-    stroke: white;
-}
-
-.downloadContainer {
-    display: flex;
-    flex-direction: row;
-    gap: var(--gap-medium);
-    justify-content: center;
-    align-items: center;
-}
-
-.saveIconContainer {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    gap: var(--gap-small);
-    cursor: pointer;
-}
-
-.saveIconContainer svg:hover {
-    fill: rgba(239, 239, 239, 0.8);
-}
-
-.saveIconContainer svg {
-    width: 20px;
-    height: 20px;
-    fill: white;
-}
-
-.saveIconContainer button {
-    background-color: transparent;
-    border: none;
-    cursor: pointer;
 }
