+++ src/features/editor/index.ts
@@ -3,10 +3,11 @@ export { default as DocumentContainer } from "./components/document-container/Do
 export { default as MenuContainer } from "./components/side-menu/menu-container/MenuContainer";
 export { default as DraggableContainer } from "./components/draggable-section-container/DraggableContainer";
 export { default as PageUtilBar } from "./components/page-util-bar/PageUtilBar";
-export { default as PageCounter } from "./components/page-util-bar/page-counter/PageCounter";
+export { default as PageCounter } from "./components/page-counter/PageCounter";
 export { default as PageButtons } from "./components/page-add-delete-buttons/PageButtons";
 export { default as SideMenu } from "./components/side-menu/SideMenu";
 export { default as TitleBar } from "./components/title-bar/TitleBar";
+export { default as SubTitleBar } from "./components/sub-title-bar/SubTitleBar";
 export { SectionConfig } from "./lib/sectionConfig";
 export { CoverLetterSectionConfig } from "./lib/coverLetterSectionConfig";
 export {
